Rename language files to fix compilation error
