+++ platformio_tasmota_env.ini
@@ -38,67 +38,67 @@ build_flags = ${common.build_flags} ${irremoteesp_full.build_flags} -DFIRMWARE_I
 build_flags = ${common.build_flags} ${irremoteesp_full.build_flags}
 
 [env:tasmota-BG]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=bg-BG
+build_flags = ${common.build_flags} -DMY_LANGUAGE=bg_BG
 
 [env:tasmota-BR]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=pt-BR
+build_flags = ${common.build_flags} -DMY_LANGUAGE=pt_BR
 
 [env:tasmota-CN]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=zh-CN
+build_flags = ${common.build_flags} -DMY_LANGUAGE=zh_CN
 
 [env:tasmota-CZ]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=cs-CZ
+build_flags = ${common.build_flags} -DMY_LANGUAGE=cs_CZ
 
 [env:tasmota-DE]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=de-DE
+build_flags = ${common.build_flags} -DMY_LANGUAGE=de_DE
 
 [env:tasmota-ES]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=es-ES
+build_flags = ${common.build_flags} -DMY_LANGUAGE=es_ES
 
 [env:tasmota-FR]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=fr-FR
+build_flags = ${common.build_flags} -DMY_LANGUAGE=fr_FR
 
 [env:tasmota-GR]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=el-GR
+build_flags = ${common.build_flags} -DMY_LANGUAGE=el_GR
 
 [env:tasmota-HE]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=he-HE
+build_flags = ${common.build_flags} -DMY_LANGUAGE=he_HE
 
 [env:tasmota-HU]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=hu-HU
+build_flags = ${common.build_flags} -DMY_LANGUAGE=hu_HU
 
 [env:tasmota-IT]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=it-IT
+build_flags = ${common.build_flags} -DMY_LANGUAGE=it_IT
 
 [env:tasmota-KO]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=ko-KO
+build_flags = ${common.build_flags} -DMY_LANGUAGE=ko_KO
 
 [env:tasmota-NL]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=nl-NL
+build_flags = ${common.build_flags} -DMY_LANGUAGE=nl_NL
 
 [env:tasmota-PL]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=pl-PL
+build_flags = ${common.build_flags} -DMY_LANGUAGE=pl_PL
 
 [env:tasmota-PT]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=pt-PT
+build_flags = ${common.build_flags} -DMY_LANGUAGE=pt_PT
 
 [env:tasmota-RO]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=ro-RO
+build_flags = ${common.build_flags} -DMY_LANGUAGE=ro_RO
 
 [env:tasmota-RU]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=ru-RU
+build_flags = ${common.build_flags} -DMY_LANGUAGE=ru_RU
 
 [env:tasmota-SE]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=sv-SE
+build_flags = ${common.build_flags} -DMY_LANGUAGE=sv_SE
 
 [env:tasmota-SK]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=sk-SK
+build_flags = ${common.build_flags} -DMY_LANGUAGE=sk_SK
 
 [env:tasmota-TR]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=tr-TR
+build_flags = ${common.build_flags} -DMY_LANGUAGE=tr_TR
 
 [env:tasmota-TW]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=zh-TW
+build_flags = ${common.build_flags} -DMY_LANGUAGE=zh_TW
 
 [env:tasmota-UK]
-build_flags = ${common.build_flags} -DMY_LANGUAGE=uk-UA
+build_flags = ${common.build_flags} -DMY_LANGUAGE=uk_UA
