+++ platformio_tasmota_env32.ini
@@ -47,88 +47,88 @@ build_flags             = ${common32.build_flags} ${irremoteesp_full.build_flags
 
 [env:tasmota32-BG]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=bg-BG
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=bg_BG
 
 [env:tasmota32-BR]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=pt-BR
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=pt_BR
 
 [env:tasmota32-CN]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=zh-CN
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=zh_CN
 
 [env:tasmota32-CZ]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=cs-CZ
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=cs_CZ
 
 [env:tasmota32-DE]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=de-DE
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=de_DE
 
 [env:tasmota32-ES]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=es-ES
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=es_ES
 
 [env:tasmota32-FR]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=fr-FR
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=fr_FR
 
 [env:tasmota32-GR]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=el-GR
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=el_GR
 
 [env:tasmota32-HE]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=he-HE
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=he_HE
 
 [env:tasmota32-HU]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=hu-HU
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=hu_HU
 
 [env:tasmota32-IT]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=it-IT
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=it_IT
 
 [env:tasmota32-KO]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=ko-KO
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=ko_KO
 
 [env:tasmota32-NL]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=nl-NL
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=nl_NL
 
 [env:tasmota32-PL]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=pl-PL
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=pl_PL
 
 [env:tasmota32-PT]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=pt-PT
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=pt_PT
 
 [env:tasmota32-RO]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=ro-RO
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=ro_RO
 
 [env:tasmota32-RU]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=ru-RU
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=ru_RU
 
 [env:tasmota32-SE]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=sv-SE
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=sv_SE
 
 [env:tasmota32-SK]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=sk-SK
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=sk_SK
 
 [env:tasmota32-TR]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=tr-TR
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=tr_TR
 
 [env:tasmota32-TW]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=zh-TW
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=zh_TW
 
 [env:tasmota32-UK]
 extends = env:tasmota32
-build_flags             = ${common32.build_flags} -DMY_LANGUAGE=uk-UA
+build_flags             = ${common32.build_flags} -DMY_LANGUAGE=uk_UA
