+++ tasmota/language/en_GB.h
