+++ tasmota/language/sk_SK.h
